Refactor/nacos-discovery-example (#2933)
refactor: nacos-discovery-example
