+++ spring-cloud-alibaba-examples/nacos-example/nacos-gateway-example/nacos-gateway-provider-example/src/main/java/com/alibaba/cloud/examples/ProviderApplication.java
@@ -19,10 +19,6 @@ package com.alibaba.cloud.examples;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 /**
  * @author xiaojing
@@ -33,21 +29,6 @@ public class ProviderApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(ProviderApplication.class, args);
-	}
-
-	@RestController
-	class EchoController {
-
-		@GetMapping("/echo/{string}")
-		public String echo(@PathVariable String string) {
-			return "hello Nacos Discovery " + string;
-		}
-
-		@GetMapping("/divide")
-		public String divide(@RequestParam Integer a, @RequestParam Integer b) {
-			return String.valueOf(a / b);
-		}
-
 	}
 
 }
